create v6

+++ src/main/java/org/example/synergy/dto/AppraisalUser.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class AppraisalUser {
     private String name;
     private String code;
-    private Integer level;
+    private Double level;
     private String organization;
     private String role;
     private Boolean hostUnit;
